Added latest updates from other branches

+++ src/components/login/LoginForm.jsx
@@ -7,31 +7,33 @@ import jwt from 'jsonwebtoken';
 import { login, setCurrentUser } from '../../actions/currentUser';
 
 class LoginForm extends React.Component {
+  
+  state = {
+    auth: getParameter('teamName') ? false : true,
+    teamId: getParameter('teamId'),
+    companyId: getParameter('companyId'),
+    email: getParameter('email'),
+    teamName: getParameter('teamName'),
+    error: false,
+    wrongConfirmPass:false
+  };
 
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      auth: getParameter('teamName') ? false : true,
-      teamName: getParameter('teamName'),
-      teamId: getParameter('teamId'),
-      email: getParameter('email'),
-      error: false
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(ev) {
+  handleChange = ev => {
     this.setState({
       [ev.target.name]: ev.target.value,
+      wrongConfirmPass: false,
       error: false
     })
   }
 
-  handleSubmit(ev) {
+  handleSubmit = (ev) => {
     ev.preventDefault();
+    const {password, passwordRepeat} = this.state
+    if (password !== passwordRepeat && !this.state.auth) {
+      this.setState({wrongConfirmPass: true})
+      setTimeout(() => {this.setState({wrongConfirmPass: false})},500)
+      return
+    }
     this.props.login(this.state).then((res) => {
       const token = res.data.jwtToken;
       localStorage.setItem('jwtToken', token);
@@ -42,8 +44,23 @@ class LoginForm extends React.Component {
   }
 
   render() {
-    const loginError = this.state.error ? "Check your login" : null;
-    const passwordError = this.state.error ? "Check your password" : null;
+    const {handleChange} = this
+    const {wrongConfirmPass} = this.state
+    const style = {}
+
+    if (wrongConfirmPass) {
+      style.confirmPassStyle = {backgroundColor: 'rgb(245,138,71)'}
+    } else {
+      style.confirmPassStyle = null
+    }
+
+    const loginError = this.state.error
+      ? "Check your login"
+      : null
+
+    const passwordError = this.state.error
+      ? "Check your password"
+      : null
 
     const auth = (
       <div className="container container__fixed">
@@ -51,20 +68,22 @@ class LoginForm extends React.Component {
           <div className="col-md-5">
             <form className="form-signin" onSubmit={this.handleSubmit}>
               <h2 className="heading">Sign In</h2>
-              <Input handleChange={this.handleChange}
-                  className="input-group input-group__grey"
-                  label="Username"
-                  error={loginError}
-                  required="true"
-                  name="username"/>
+              <Input
+                handleChange={handleChange}
+                className="input-group input-group__grey"
+                label="Username"
+                error={loginError}
+                required="true"
+                name="username"/>
 
-              <Input handleChange={this.handleChange}
-                  type="password"
-                  error={passwordError}
-                  className="input-group input-group__grey"
-                  label="Password"
-                  required="true"
-                  name="password"/>
+              <Input
+                handleChange={handleChange}
+                type="password"
+                error={passwordError}
+                className="input-group input-group__grey"
+                label="Password"
+                required="true"
+                name="password"/>
 
               <Checkbox label="Remember me?" name="rememberme"/>
 
@@ -80,29 +99,34 @@ class LoginForm extends React.Component {
         <div className="row center-md">
           <div className="col-md-5">
             <form className="form-signup" onSubmit={this.handleSubmit}>
-              <h2 className="form-signin-heading">Signing Up to { this.state.teamName } </h2>
-              <input type="hidden" name="email" defaultValue={ this.state.email }/>
-              <input type="hidden" name="teamId" defaultValue={ this.state.teamId }/>
+              <h2 className="form-signin-heading">Signing Up to {this.state.teamName}
+              </h2>
+              <input type="hidden" name="email" defaultValue={this.state.email}/>
+              <input type="hidden" name="teamId" defaultValue={this.state.teamId}/>
 
-              <Input handleChange={this.handleChange}
-                  className="input-group input-group__grey"
-                  label="Login"
-                  required="true"
-                  name="login"/>
-              <Input handleChange={this.handleChange}
-                  className="input-group input-group__grey"
-                  type="password"
-                  label="Password"
-                  required="true"
-                  name="password"/>
-
-              <Input handleChange={this.handleChange}
+              <Input
+                handleChange={handleChange}
+                className="input-group input-group__grey"
+                label="Login"
+                required="true"
+                name="login"/>
+              <Input
+                handleChange={handleChange}
+                className="input-group input-group__grey"
+                type="password"
+                label="Password"
+                required="true"
+                name="password"/>
+              <div style={style.confirmPassStyle}>
+                <Input
+                  handleChange={handleChange}
                   className="input-group input-group__grey"
                   type="password"
                   label="Repeat Password"
                   required="true"
-                  name="password-repeat"/>
-
+                  name="passwordRepeat"
+                  />
+              </div>
               <button className="btn btn__lg btn__blue signup_btn" type="submit">Sign up</button>
             </form>
           </div>
@@ -111,7 +135,9 @@ class LoginForm extends React.Component {
     );
 
     return (
-      <div className="login-form-container">{ this.state.auth ? auth : reg }</div>
+      <div className="login-form-container">
+        { this.state.auth ? auth : reg }
+      </div>
     )
   }
 }
